Update index.html
Step 8 Passed

HTML attributes are special words used inside the opening tag of an element to control the element's behavior. The src attribute in an img element specifies the image's URL (where the image is located).

Here is an example of an img element with a src attribute pointing to the freeCodeCamp logo:

<img src="https://cdn.freecodecamp.org/platform/universal/fcc_secondary.svg">

Inside the existing img element, add an src attribute with this URL:

https://cdn.freecodecamp.org/curriculum/cat-photo-app/relaxing-cat.jpg