Update index.html
Step 9 Passed

All img elements should have an alt attribute. The alt attribute's text is used for screen readers to improve accessibility and is displayed if the image fails to load. For example, <img src="cat.jpg" alt="A cat"> has an alt attribute with the text A cat.

Inside the img element, add an alt attribute with this text:

A cute orange cat lying on its back