Update index.html
Step 11 Passed

A link's text must be placed between the opening and closing tags of an anchor (a) element. For example, <a href="https://www.freecodecamp.org">click here to go to freeCodeCamp.org</a> is a link with the text click here to go to freeCodeCamp.org.

Add the anchor text link to cat pictures to the anchor element. This will become the link's text.