Update index.html
Step 69 Passed

You can set browser behavior by adding self-closing meta elements in the head. Here's an example:

<meta attribute="value">

Inside the head element, nest a meta element with an attribute named charset. Set to the value to utf-8 which tells the browser how to encode characters for the page.

Note that meta elements are self-closing.